requirements.txt
beautifulsoup4==4.9.3
requests==2.25.1
uvicorn==0.30.1
pymongo==4.7.3
pytesseract==0.3.10
pydantic==2.7.3
pillow==10.3.0
fastapi==0.111.0
certifi==2024.6.2